+++ src/System.CommandLine.Tests/CompilationTests.cs
@@ -8,17 +8,18 @@ using System.CommandLine.Tests.Utility;
 using System.IO;
 using System.Text;
 using FluentAssertions;
+using Microsoft.DotNet.PlatformAbstractions;
 using Xunit;
 using Xunit.Abstractions;
 
 namespace System.CommandLine.Tests;
 
-public class TrimmingTests
+public class CompilationTests
 {
     private readonly ITestOutputHelper _output;
     private readonly string _systemCommandLineDllPath;
 
-    public TrimmingTests(ITestOutputHelper output)
+    public CompilationTests(ITestOutputHelper output)
     {
         _output = output;
 
@@ -29,25 +30,35 @@ public class TrimmingTests
     [InlineData("")]
     [InlineData("-p:PublishSingleFile=true")]
     public void App_referencing_system_commandline_can_be_trimmed(string additionalArgs)
+        => PublishAndValidate("Trimming", "warning IL", additionalArgs);
+
+    [ReleaseBuildOnlyFact]
+    public void App_referencing_system_commandline_can_be_compiled_ahead_of_time()
+        => PublishAndValidate("NativeAOT", "AOT analysis warning");
+
+    private void PublishAndValidate(string appName, string warningText, string additionalArgs = null)
     {
         var stdOut = new StringBuilder();
         var stdErr = new StringBuilder();
 
-        var workingDirectory = Path.Combine(Directory.GetCurrentDirectory(), "TestApps", "Trimming");
-        
+        var workingDirectory = Path.Combine(Directory.GetCurrentDirectory(), "TestApps", appName);
+
+        string rId = GetPortableRuntimeIdentifier();
+
         Process.RunToCompletion(
             DotnetMuxer.Path.FullName,
-            "clean -c Release -r win-x64",
+            $"clean -c Release -r {rId}",
             workingDirectory: workingDirectory);
 
-        var commandLine = string.Format(
-            "publish -c Release -r win-x64 --self-contained -p:SystemCommandLineDllPath=\"{0}\" -p:TreatWarningsAsErrors=true -p:PublishTrimmed=true {1}",
+        string publishCommand = string.Format(
+            "publish -c Release -r {0} --self-contained -p:SystemCommandLineDllPath=\"{1}\" -p:TreatWarningsAsErrors=true {2}",
+            rId,
             _systemCommandLineDllPath,
             additionalArgs);
 
         var exitCode = Process.RunToCompletion(
             DotnetMuxer.Path.FullName,
-            commandLine,
+            publishCommand,
             s =>
             {
                 _output.WriteLine(s);
@@ -60,9 +71,15 @@ public class TrimmingTests
             },
             workingDirectory);
 
-        stdOut.ToString().Should().NotContain("warning IL");
+        stdOut.ToString().Should().NotContain(warningText);
         stdErr.ToString().Should().BeEmpty();
         exitCode.Should().Be(0);
+    }
+
+    private static string GetPortableRuntimeIdentifier()
+    {
+        string osPart = OperatingSystem.IsWindows() ? "win" : (OperatingSystem.IsMacOS() ? "osx" : "linux");
+        return $"{osPart}-{RuntimeEnvironment.RuntimeArchitecture}";
     }
 }
